+++ app.py
@@ -418,7 +418,7 @@ def app_sendonly():
 
 if __name__ == "__main__":
     logging.basicConfig(
-        format="[%(asctime)s] %(levelname)7s from %(name)s in %(filename)s:%(lineno)d: "
+        format="[%(asctime)s] %(levelname)7s from %(name)s in %(pathname)s:%(lineno)d: "
         "%(message)s",
         force=True,
     )
@@ -427,5 +427,8 @@ if __name__ == "__main__":
 
     st_webrtc_logger = logging.getLogger("streamlit_webrtc")
     st_webrtc_logger.setLevel(logging.DEBUG)
+
+    fsevents_logger = logging.getLogger("fsevents")
+    fsevents_logger.setLevel(logging.WARNING)
 
     main()