+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Fix
 - Internal fix, #126.
+- Set log level to `fsevents` logger, whose logs have been noisy, #129.
 
 ## [0.8.0] - 2021-04-09
 ### Fix